# Hashed utility for ValueMap to improve hashing performance