@fraillt 2024-11-13
> @fraillt This PR is trying to bring two changes at once.
> 
> Could you please split this into two different PRs?
> 
>     1. One PR which only has changes related to caching the computed hash value
> 
>     2. Another one which uses a different Hasher (could you try [aHash](https://github.com/tkaitchuck/aHash)?). Also, please note this would have to be feature flagged. We do not want to use a custom hasher by default.

Updated PR, and also task description.
I'll create another PR, with different hasher functions once/if this lands.