@cijothomas 2024-11-14
Ran stress/benchmarks
Benchmarks - within noise.
Stress test metrics_counter - Drops from 14.5M/sec to 11M/sec