@cijothomas 2024-11-14
I see this will have huge value if we were to implement sharing ourselves, so caching the hash value will definitely help. Before proceeding further, I'd like to suggest to confirm that sharing ourselves is the right direction. See https://github.com/open-telemetry/opentelemetry-rust/pull/2288#discussion_r1841500901