@fraillt 2024-11-14
> Benchmarks - within noise.(Overflow scenario does improve due to avoiding the need for hash calc multiple times)
> Stress test metrics_counter - Drops from 14.5M/sec to 11M/sec

Yes, benchmarks are basically within the noise, but it's technically an improvement, this revision avoids extra hashing in HashMap, because of `HashedNoOpBuilder` on hashmap. 

Regarding stress they are so weird and sensitive on so many factors... on my current machine, I consistently observe opposite results, maybe ~10% improvement, on 1.81 rust version. but your machine is totally different, -20% performance degradation... wow :)
I have my old SBC (single-board-computer) with ARM processor, maybe I'll need to wipe off the dust, and use it for some stress testing :)